renamed files of sparsityPattern and linearSystem
