+++ include/NeoFOAM/finiteVolume/cellCentred/operators/sparsityPattern.hpp
