+++ src/CMakeLists.txt
@@ -38,7 +38,7 @@ target_sources(
           "finiteVolume/cellCentred/boundary/boundary.cpp"
           "finiteVolume/cellCentred/operators/gaussGreenGrad.cpp"
           "finiteVolume/cellCentred/operators/gaussGreenDiv.cpp"
-          "finiteVolume/cellCentred/operators/fvccSparsityPattern.cpp"
+          "finiteVolume/cellCentred/operators/sparsityPattern.cpp"
           "finiteVolume/cellCentred/interpolation/linear.cpp"
           "finiteVolume/cellCentred/interpolation/upwind.cpp"
           "timeIntegration/timeIntegration.cpp"
